# tsdb: use simpler map key to improve exemplar ingest performance