--- a/tsdb/exemplar.go
+++ b/tsdb/exemplar.go
@@ -202,7 +202,8 @@ Outer:
 }
 
 func (ce *CircularExemplarStorage) ValidateExemplar(l labels.Labels, e exemplar.Exemplar) error {
-	seriesLabels := l.String()
+	var buf [1024]byte
+	seriesLabels := l.Bytes(buf[:])
 
 	// TODO(bwplotka): This lock can lock all scrapers, there might high contention on this on scale.
 	// Optimize by moving the lock to be per series (& benchmark it).
@@ -213,7 +214,7 @@ func (ce *CircularExemplarStorage) ValidateExemplar(l labels.Labels, e exemplar.
 
 // Not thread safe. The append parameters tells us whether this is an external validation, or internal
 // as a result of an AddExemplar call, in which case we should update any relevant metrics.
-func (ce *CircularExemplarStorage) validateExemplar(l string, e exemplar.Exemplar, append bool) error {
+func (ce *CircularExemplarStorage) validateExemplar(key []byte, e exemplar.Exemplar, append bool) error {
 	if len(ce.exemplars) <= 0 {
 		return storage.ErrExemplarsDisabled
 	}
@@ -230,7 +231,7 @@ func (ce *CircularExemplarStorage) validateExemplar(l string, e exemplar.Exempla
 		}
 	}
 
-	idx, ok := ce.index[l]
+	idx, ok := ce.index[string(key)]
 	if !ok {
 		return nil
 	}
@@ -305,13 +306,14 @@ func (ce *CircularExemplarStorage) Resize(l int64) int {
 // migrate is like AddExemplar but reuses existing structs. Expected to be called in batch and requires
 // external lock and does not compute metrics.
 func (ce *CircularExemplarStorage) migrate(entry *circularBufferEntry) {
-	seriesLabels := entry.ref.seriesLabels.String()
+	var buf [1024]byte
+	seriesLabels := entry.ref.seriesLabels.Bytes(buf[:])
 
-	idx, ok := ce.index[seriesLabels]
+	idx, ok := ce.index[string(seriesLabels)]
 	if !ok {
 		idx = entry.ref
 		idx.oldest = ce.nextIndex
-		ce.index[seriesLabels] = idx
+		ce.index[string(seriesLabels)] = idx
 	} else {
 		entry.ref = idx
 		ce.exemplars[idx.newest].next = ce.nextIndex
@@ -329,7 +331,8 @@ func (ce *CircularExemplarStorage) AddExemplar(l labels.Labels, e exemplar.Exemp
 		return storage.ErrExemplarsDisabled
 	}
 
-	seriesLabels := l.String()
+	var buf [1024]byte
+	seriesLabels := l.Bytes(buf[:])
 
 	// TODO(bwplotka): This lock can lock all scrapers, there might high contention on this on scale.
 	// Optimize by moving the lock to be per series (& benchmark it).
@@ -345,11 +348,11 @@ func (ce *CircularExemplarStorage) AddExemplar(l labels.Labels, e exemplar.Exemp
 		return err
 	}
 
-	_, ok := ce.index[seriesLabels]
+	_, ok := ce.index[string(seriesLabels)]
 	if !ok {
-		ce.index[seriesLabels] = &indexEntry{oldest: ce.nextIndex, seriesLabels: l}
+		ce.index[string(seriesLabels)] = &indexEntry{oldest: ce.nextIndex, seriesLabels: l}
 	} else {
-		ce.exemplars[ce.index[seriesLabels].newest].next = ce.nextIndex
+		ce.exemplars[ce.index[string(seriesLabels)].newest].next = ce.nextIndex
 	}
 
 	if prev := ce.exemplars[ce.nextIndex]; prev == nil {
@@ -357,12 +360,13 @@ func (ce *CircularExemplarStorage) AddExemplar(l labels.Labels, e exemplar.Exemp
 	} else {
 		// There exists exemplar already on this ce.nextIndex entry, drop it, to make place
 		// for others.
-		prevLabels := prev.ref.seriesLabels.String()
+		var buf [1024]byte
+		prevLabels := prev.ref.seriesLabels.Bytes(buf[:])
 		if prev.next == noExemplar {
 			// Last item for this series, remove index entry.
-			delete(ce.index, prevLabels)
+			delete(ce.index, string(prevLabels))
 		} else {
-			ce.index[prevLabels].oldest = prev.next
+			ce.index[string(prevLabels)].oldest = prev.next
 		}
 	}
 
@@ -370,8 +374,8 @@ func (ce *CircularExemplarStorage) AddExemplar(l labels.Labels, e exemplar.Exemp
 	// since this is the first exemplar stored for this series.
 	ce.exemplars[ce.nextIndex].next = noExemplar
 	ce.exemplars[ce.nextIndex].exemplar = e
-	ce.exemplars[ce.nextIndex].ref = ce.index[seriesLabels]
-	ce.index[seriesLabels].newest = ce.nextIndex
+	ce.exemplars[ce.nextIndex].ref = ce.index[string(seriesLabels)]
+	ce.index[string(seriesLabels)].newest = ce.nextIndex
 
 	ce.nextIndex = (ce.nextIndex + 1) % len(ce.exemplars)