--- a/tsdb/exemplar_test.go
+++ b/tsdb/exemplar_test.go
@@ -492,18 +492,23 @@ func BenchmarkAddExemplar(b *testing.B) {
 
 	for _, n := range []int{10000, 100000, 1000000} {
 		b.Run(fmt.Sprintf("%d", n), func(b *testing.B) {
-			exs, err := NewCircularExemplarStorage(int64(n), eMetrics)
-			require.NoError(b, err)
-			es := exs.(*CircularExemplarStorage)
-
-			b.ResetTimer()
-			l := labels.Labels{{Name: "service", Value: strconv.Itoa(0)}}
-			for i := 0; i < n; i++ {
-				if i%100 == 0 {
-					l = labels.Labels{{Name: "service", Value: strconv.Itoa(i)}}
-				}
-				err = es.AddExemplar(l, exemplar.Exemplar{Value: float64(i), Ts: int64(i), Labels: exLabels})
+			for j := 0; j < b.N; j++ {
+				b.StopTimer()
+				exs, err := NewCircularExemplarStorage(int64(n), eMetrics)
 				require.NoError(b, err)
+				es := exs.(*CircularExemplarStorage)
+				l := labels.Labels{{Name: "service", Value: strconv.Itoa(0)}}
+				b.StartTimer()
+
+				for i := 0; i < n; i++ {
+					if i%100 == 0 {
+						l = labels.Labels{{Name: "service", Value: strconv.Itoa(i)}}
+					}
+					err = es.AddExemplar(l, exemplar.Exemplar{Value: float64(i), Ts: int64(i), Labels: exLabels})
+					if err != nil {
+						require.NoError(b, err)
+					}
+				}
 			}
 		})
 	}
@@ -543,24 +548,24 @@ func BenchmarkResizeExemplars(b *testing.B) {
 	}
 
 	for _, tc := range testCases {
-		exs, err := NewCircularExemplarStorage(tc.startSize, eMetrics)
-		require.NoError(b, err)
-		es := exs.(*CircularExemplarStorage)
+		b.Run(fmt.Sprintf("%s-%d-to-%d", tc.name, tc.startSize, tc.endSize), func(b *testing.B) {
+			for j := 0; j < b.N; j++ {
+				b.StopTimer()
+				exs, err := NewCircularExemplarStorage(tc.startSize, eMetrics)
+				require.NoError(b, err)
+				es := exs.(*CircularExemplarStorage)
 
-		for i := 0; i < int(float64(tc.startSize)*float64(1.5)); i++ {
-			l := labels.FromStrings("service", strconv.Itoa(i))
+				for i := 0; i < int(float64(tc.startSize)*float64(1.5)); i++ {
+					l := labels.FromStrings("service", strconv.Itoa(i))
 
-			err = es.AddExemplar(l, exemplar.Exemplar{Value: float64(i), Ts: int64(i)})
-			require.NoError(b, err)
-		}
-		saveIndex := es.index
-		saveExemplars := es.exemplars
-
-		b.Run(fmt.Sprintf("%s-%d-to-%d", tc.name, tc.startSize, tc.endSize), func(t *testing.B) {
-			es.index = saveIndex
-			es.exemplars = saveExemplars
-			b.ResetTimer()
-			es.Resize(tc.endSize)
+					err = es.AddExemplar(l, exemplar.Exemplar{Value: float64(i), Ts: int64(i)})
+					if err != nil {
+						require.NoError(b, err)
+					}
+				}
+				b.StartTimer()
+				es.Resize(tc.endSize)
+			}
 		})
 	}
 }