Prometheus holds an index of exemplars so it can discard the oldest one for a series when a new one is added.
Since the keys are not for human eyes, we can use a simpler format and save the effort of quoting label values.

Very similar to #9697

The buffer on the stack is to reduce heap allocations; I gave it the arbitrary size of 1024, which is enough for 20 label/value pairs where each string is 24 bytes long.

I also had to fix the benchmarks, which were reporting results like like 0.01 ns/op, which is nonsense.
Go benchmarks are expected to do an amount of work that varies with the `b.N` parameter.
After fixing:
```
name                                        old time/op  new time/op  delta
AddExemplar/10000-4                         6.72ms ± 4%  4.19ms ± 3%  -37.65%  (p=0.008 n=5+5)
AddExemplar/100000-4                        84.2ms ±48%  40.9ms ± 2%  -51.50%  (p=0.016 n=5+4)
AddExemplar/1000000-4                        676ms ± 2%   394ms ± 3%  -41.68%  (p=0.008 n=5+5)
ResizeExemplars/grow-100000-to-200000-4     55.6ms ±28%  35.0ms ±31%  -37.05%  (p=0.008 n=5+5)
ResizeExemplars/shrink-100000-to-50000-4    24.3ms ± 3%  20.3ms ±55%     ~     (p=0.190 n=4+5)
ResizeExemplars/grow-1000000-to-2000000-4    790ms ±19%   481ms ±13%  -39.13%  (p=0.008 n=5+5)
ResizeExemplars/shrink-1000000-to-500000-4   357ms ±15%   189ms ±18%  -47.23%  (p=0.008 n=5+5)
```