@bboreham 2022-01-05
After more profiling, I pushed an update improving the benchmark again and adding an initial size to the map.
Now it checks `err != nil` before calling `require.NoError` in tight loops, because the latter involves a stack walk which is about as expensive as the operation we are timing.